requirements.txt
fastapi~=0.111.0
uvicorn==0.30.1
openai==1.55.3
langchain-openai==0.1.25
langchain~=0.2.2
pydantic==2.9.2
pydantic_settings~=2.3.1
urllib3~=2.2.1
requests~=2.32.3
python-dotenv
python-multipart